Code with preliminary data
When patient ID was unavailable, we used this code to filter out all ECG's whose date was the same as any ajmaline test (independently of the patient ID).
This processing doesn't do filter correction.
The result is 9 models tuned on the validation set. 4 models were preselected as candidates to the final ECG model.